+++ hana/lib/HANAService.js
@@ -98,7 +98,12 @@ class HANAService extends SQLService {
     this.ensureDBC().set(variables)
   }
 
-  async onSELECT({ query, data }) {
+  async onSELECT(req) {
+    const { query, data } = req
+    if (!query.target || query.target._unresolved) {
+      return super.onSELECT(req)
+    }
+
     // REVISIT: disable this for queries like (SELECT 1)
     // Will return multiple rows with objects inside
     query.SELECT.expand = 'root'
@@ -119,6 +124,7 @@ class HANAService extends SQLService {
   }
 
   async onINSERT({ query, data }) {
+    try {
     const { sql, entries, cqn } = this.cqn2sql(query, data)
     if (!sql) return // Do nothing when there is nothing to be done
     const ps = await this.prepare(sql)
@@ -129,6 +135,17 @@ class HANAService extends SQLService {
         : ps.run(entries[0])
       : ps.run())
     return new this.class.InsertResults(cqn, results)
+    } catch (err) {
+      throw _not_unique(err, 'ENTITY_ALREADY_EXISTS')
+    }
+  }
+
+  async onUPDATE(req) {
+    try {
+      return await super.onUPDATE(req)
+    } catch (err) {
+      throw _not_unique(err, 'UNIQUE_CONSTRAINT_VIOLATION') || err
+    }
   }
 
   // Allow for running complex expand queries in a single statement
@@ -249,7 +266,7 @@ class HANAService extends SQLService {
       this.temporary = this.temporary || []
       this.temporaryValues = this.temporaryValues || []
 
-      const { limit, one, orderBy, expand, columns, localized, count, parent } = q.SELECT
+      const { limit, one, orderBy, expand, columns = ['*'], localized, count, parent } = q.SELECT
 
       const walkAlias = q => {
         if (q.args) return q.as || walkAlias(q.args[0])
@@ -616,7 +633,7 @@ class HANAService extends SQLService {
       const keyCompare =
         keys &&
         Object.keys(keys)
-          .filter(k => !keys[k].isAssociation)
+          .filter(k => !keys[k].isAssociation && !keys[k].virtual)
           .map(k => `NEW.${this.quote(k)}=OLD.${this.quote(k)}`)
           .join(' AND ')
 
@@ -647,7 +664,9 @@ class HANAService extends SQLService {
     }
 
     where(xpr) {
-      return this.xpr({ xpr: [...xpr, 'THEN'] }).slice(0, -4)
+      xpr = { xpr }
+      const suffix = this.is_comparator(xpr) ? '' : ' = TRUE'
+      return `${this.xpr(xpr)}${suffix}`
     }
 
     having(xpr) {
@@ -673,6 +692,7 @@ class HANAService extends SQLService {
         */
       }
 
+      let endWithCompare = false
       if (!_internal) {
         for (let i = 0; i < xpr.length; i++) {
           let x = xpr[i]
@@ -683,6 +703,7 @@ class HANAService extends SQLService {
             // HANA does not support comparators in all clauses (e.g. SELECT 1>0 FROM DUMMY)
             // HANA does not have an 'IS' or 'IS NOT' operator
             if (x in compareOperators) {
+              endWithCompare = true
               const left = xpr[i - 1]
               const right = xpr[i + 1]
               const ifNull = compareOperators[x]
@@ -725,16 +746,25 @@ class HANAService extends SQLService {
       for (let i = 0; i < xpr.length; ++i) {
         const x = xpr[i]
         if (typeof x === 'string') {
+          const up = x.toUpperCase()
+          if (up in logicOperators) {
+            // Force current expression to end with a comparison
+            endWithCompare = true
+          }
+          if (endWithCompare && (up in caseOperators || up === ')')) {
+            endWithCompare = false
+          }
           sql.push(this.operator(x, i, xpr))
         } else if (x.xpr) sql.push(`(${this.xpr(x, caseSuffix)})`)
         // default
         else sql.push(this.expr(x))
       }
 
-      // HANA does not allow WHERE TRUE so when the expression is only a single entry "= TRUE" is appended
-      if (caseSuffix && (
-        xpr.length === 1 || xpr.at(-1) === '')) {
-        sql.push(caseSuffix)
+      if (endWithCompare) {
+        const suffix = this.operator('OR', xpr.length, xpr).slice(0, -3)
+        if (suffix) {
+          sql.push(suffix)
+        }
       }
 
       return `${sql.join(' ')}`
@@ -765,13 +795,14 @@ class HANAService extends SQLService {
      * @returns
      */
     is_comparator({ xpr }, start) {
+      const local = start != null
       for (let i = start ?? xpr.length; i > -1; i--) {
         const cur = xpr[i]
         if (cur == null) continue
         if (typeof cur === 'string') {
           const up = cur.toUpperCase()
           // When a compare operator is found the expression is a comparison
-          if (up in compareOperators) return true
+          if (up in compareOperators || (!local && up in logicOperators)) return true
           // When a case operator is found it is the start of the expression
           if (up in caseOperators) break
           continue
@@ -781,17 +812,17 @@ class HANAService extends SQLService {
       return false
     }
 
-    list(list) {
-      const first = list.list[0]
-      // If the list only contains of lists it is replaced with a json function and a placeholder
-      if (this.values && first.list && !first.list.find(v => !v.val)) {
-        const extraction = first.list.map((v, i) => `"${i}" ${this.constructor.InsertTypeMap[typeof v.val]()} PATH '$.V${i}'`)
-        this.values.push(JSON.stringify(list.list.map(l => l.list.reduce((l, c, i) => { l[`V${i}`] = c.val; return l }, {}))))
-        return `(SELECT * FROM JSON_TABLE(?, '$' COLUMNS(${extraction})))`
-      }
-      // Call super for normal SQL behavior
-      return super.list(list)
-    }
+        list(list) {
+          const first = list.list[0]
+          // If the list only contains of lists it is replaced with a json function and a placeholder
+          if (this.values && first.list && !first.list.find(v => !v.val)) {
+            const extraction = first.list.map((v, i) => `"${i}" ${this.constructor.InsertTypeMap[typeof v.val]()} PATH '$.V${i}'`)
+            this.values.push(JSON.stringify(list.list.map(l => l.list.reduce((l, c, i) => { l[`V${i}`] = c.val; return l }, {}))))
+            return `(SELECT * FROM JSON_TABLE(?, '$' COLUMNS(${extraction})))`
+          }
+          // Call super for normal SQL behavior
+          return super.list(list)
+        }
 
     quote(s) {
       // REVISIT: casing in quotes when reading from entities it uppercase
@@ -821,10 +852,13 @@ class HANAService extends SQLService {
       const requiredColumns = !elements
         ? []
         : Object.keys(elements)
-          .filter(
-            e =>
-              (elements[e]?.[annotation] || (!isUpdate && elements[e]?.default)) && !columns.find(c => c.name === e),
-          )
+          .filter(e => {
+            if (elements[e]?.virtual) return false
+            if (columns.find(c => c.name === e)) return false
+            if (elements[e]?.[annotation]) return true
+            if (!isUpdate && elements[e]?.default) return true
+            return false
+          })
           .map(name => ({ name, sql: 'NULL' }))
 
       const keyZero = this.quote(
@@ -1066,6 +1100,16 @@ const createContainerTenant = fs.readFileSync(path.resolve(__dirname, 'scripts/c
 
 Buffer.prototype.toJSON = function () {
   return this.toString('hex')
+}
+
+function _not_unique(err, code) {
+  if (err.code === 301)
+    return Object.assign(err, {
+      originalMessage: err.message, // FIXME: required because of next line
+      message: code, // FIXME: misusing message as code
+      code: 400, // FIXME: misusing code as (http) status
+    })
+  return err
 }
 
 const is_regexp = x => x?.constructor?.name === 'RegExp' // NOTE: x instanceof RegExp doesn't work in repl